fix: print legend format fix: styling in legend
refactor(PrintMapFormat): improves DOM selection for legend to print from parent-parent-sibling to a class selector that gets added in layersControl. Also changes it so the legend is created from each item, as opposed to group, so there will be no gaps and it goes left to right up to down.
refactor(NWS layers): changes fallback color from white to black
fix(layersControl): fixes padding and margin issues. Fixes component/functional nesting which was causing errors in the PrintMapFormat